bilibili_video_agent/requirements.txt
langchain>=0.2.14
langchain-community>=0.2.12
langchain-google-genai>=2.0.1
httpx>=0.25.1
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
langgraph>=0.2.74